SpiDy.jl documentation

SpiDy.jl documentation

Noise

Spectral density

Stochastic field

Coupling function

Dynamics

Index

Noise

SpiDy.spectrum — Method

spectrum(n::ClassicalNoise)

Returns the classical noise at temperature n.T as a function of ω.

SpiDy.spectrum — Method

spectrum(n::QuantumNoise)

Returns the quantum noise at temperature n.T as a function of ω.

SpiDy.spectrum — Method

spectrum(n::NoZeroQuantumNoise)

Returns the quantum noise at temperature n.T as a function of ω. It differs from spectrum(n::QuantumNoise) in the fact that the zero-point fluctuations are removed.

Spectral density

SpiDy.sd — Method

sd(J::GenericSD)

Spectral density for generic shapes.

SpiDy.sdoverω — Method

sdoverω(J::GenericSD)

Spectral-density-divided-by-ω for generic shapes.

SpiDy.sdoverω — Method

sdoverω(J::LorentzianSD)

Spectral density divided by ω which naturally defines sd(J::LorentzianSD).

SpiDy.reorgenergy — Method

reorgenergy(J::GenericSD)

Reorganization energy numerically integrated as $\int_0^\infty \text{sdoverω}(\omega)d\omega$.

SpiDy.reorgenergy — Method

reorgenergy(J::LorentzianSD)

Analytical reorganization energy for Lorentzian spectral density.

SpiDy.kernel — Method

kernel(J::LorentzianSD)

Specific damping kernel for a Lorentzian spectral density defined by the parameters in J. It returns a function depending on ω.

SpiDy.psd — Method

psd(J::GenericSD, noise::Noise)

Power spectral density depending on parameters J and noise. It returns a function of ω.

SpiDy.psd — Method

psd(J::LorentzianSD, noise::ClassicalNoise)

Returns the analytical expression for power spectrum depending on Lorentzian spectral density and Classical noise.

Stochastic field

SpiDy.bfield — Function

bfield(N, Δt, J::GenericSD, noise::Noise, distro=nothing)

Returns the stochastic field b(t). It is evaluated using the Lorentzian spectral density defined by the parameters J, the classical/quantum/quantum-no-zero-energy noise. The sampling of the stochastic noise is done in frequency space. The default stochastic noise is white noise having Gaussian distribution but different distributions can be specified. N defines the number of steps and Δt defines the time step.

<!– ## Spin state

Coupling function

Dynamics

Missing docstring.

Missing docstring for diffeqsolver(N, Δt, J::GenericSD, noise::Noise, distro=nothing). Check Documenter's build log for details.

–>

Index

SpiDy.bfield

SpiDy.kernel

SpiDy.psd

SpiDy.psd

SpiDy.reorgenergy

SpiDy.reorgenergy

SpiDy.sd

SpiDy.sdoverω

SpiDy.sdoverω

SpiDy.spectrum

SpiDy.spectrum

SpiDy.spectrum