Noise

SpiDy.Noise — Type

abstract type Noise

An abstract type used to represent different kinds of environment noise.

Any user-defined subtype of Noise should implement a method spectrum(::NoiseSubType, ω) which returns the spectrum of the noise at the given frequency ω.

SpiDy.ClassicalNoise — Type

struct ClassicalNoise{TT<:Real} <: Noise

A subtype of Noise used to represent classical noise, with spectrum

\[\mathcal{S}(\omega) = \frac{2k_\mathrm{B}T}{\hbar\omega}\]

Fields

T::TT: The temperature of the environment associated with the noise.

SpiDy.QuantumNoise — Type

struct QuantumNoise{TT<:Real} <: Noise

A subtype of Noise used to represent quantum noise, with spectrum

\[\mathcal{S}(\omega) = \coth\left(\frac{\hbar\omega}{2k_\mathrm{B}T}\right)\]

Fields

T::TT: The temperature of the environment associated with the noise.

SpiDy.NoZeroQuantumNoise — Type

struct NoZeroQuantumNoise{TT<:Real} <: Noise

A subtype of Noise used to represent quantum noise with zero-point-fluctuations removed, that is with spectrum

\[\mathcal{S}(\omega) = \coth\left(\frac{\hbar\omega}{2k_\mathrm{B}T}\right) - 1\]

Fields

T::TT: The temperature of the environment associated with the noise.

SpiDy.spectrum — Function

spectrum(::Noise)

Return the spectrum of the given noise as a function of frequency.

Arguments

n::Noise: The noise type.

Returns

A function that takes a frequency ω and returns the corresponding spectrum value.

SpiDy.psd — Function

function psd(J::AbstractSD, noise::Noise)

Calculate the Power Spectral Density (PSD) for a given environment spectral density J and noise model noise.

Arguments

J::AbstractSD: The environment spectral density.

noise::Noise: The noise model for the environment.

Note: The LorentzianSD type is provided by the SpectralDensities.jl package.

Returns

A function psd(ω) that calculates the PSD at a given frequency ω.