Spectral density

SpiDy.GenericSD — Type

GenericSD

Definition of the abstract type GenericSD.

SpiDy.LorentzianSD — Type

LorentzianSD{T<:Real}

Returns a LorentzianSD structure of type GenericSD built by passing 3 Real values. The values are ordered as LorentzianSD(α, ω0, Γ).

Examples

julia> LorentzianSD(1., 3., 8.)

SpiDy.PolySD — Type

PolySD{T<:Real}

Returns a PolySD structure of type GenericSD built by passing 3 Real values. The values are ordered as PolySD(s, α, ωcut).

Examples

julia> PolySD(1., 3., 100.)

SpiDy.sd — Method

sd(J::GenericSD)

Defines the spectral density for generic shapes GenericSD. The returned function depends on ω.

SpiDy.sdoverω — Method

sdoverω(J::GenericSD)

Returns the spectral density divided by ω for generic shapes GenericSD. The returned function depends on ω.

SpiDy.sdoverω — Method

sdoverω(J::LorentzianSD)

Returns the spectral density divided by ω for LorentzianSD shapes which naturally defines sd(J::LorentzianSD). The returned function depends on ω.

SpiDy.sd — Method

sdoverω(J::PolySD)

Returns the spectral density for PolySD shapes with exponential cut-off which naturally defines sdoverω(J::PolySD). The returned function depends on ω.

SpiDy.reorgenergy — Method

reorgenergy(J::GenericSD)

Returns the reorganization energy numerically integrated as $\int_0^\infty \text{sdoverω}(\omega)d\omega$.

SpiDy.reorgenergy — Method

reorgenergy(J::LorentzianSD)

Returns the analytical reorganization energy for LorentzianSD shapes.

SpiDy.kernel — Method

kernel(J::LorentzianSD)

Returns the specific damping kernel for a Lorentzian spectral density defined by the parameters in J. The returned function depends on ω.

SpiDy.imagkernel — Method

imagkernel(J::GenericSD)

Returns the imaginary part of the damping kernel for a Generic spectral density real and anti-symmetric in ω. Of this kind, we find Lorentzian spectral densities, and Polynomial spectral densities with $s\in\mathbb{N}$. The spectral density is defined by the parameters in J. The returned function depends on ω.

SpiDy.psd — Method

psd(J::GenericSD, noise::Noise)

Returns the power spectral density depending on parameters J and noise. The returned function depends on ω.

SpiDy.psd — Method

psd(J::LorentzianSD, noise::ClassicalNoise)

Returns the analytical expression for power spectrum depending on Lorentzian spectral density and Classical noise. The returned function depends on ω.