Stochastic field

SpiDy.bfield — Function

bfield(N, Δt, J::AbstractSD, noise::Noise; distro=Normal(0., 1/sqrt(Δt)), interpolation=true)

Generate a stochastic field realisation time series (of length N and spacing Δt) based on the given noise model noise and spectral density J.

Arguments

N: The number of time steps.

Δt: The time step size.

J::AbstractSD: The environment spectral density.

noise::Noise: The noise model for the environment.

distro=Normal(0., 1/sqrt(Δt)): (Optional) The distribution of noise samples. Default is a normal distribution with mean 0 and standard deviation 1/sqrt(Δt).

interpolation=true: (Optional) Specifies whether to use linear interpolation for the stochastic field time series. Default is true.

Note: The AbstractSD type is provided by the SpectralDensities.jl package.

Returns

A time series of the stochastic field values.