SpiDy.jl documentation

SpiDy.jl documentation

Noise

Spectral density

Stochastic field

Coupling tensor

Dynamics

Index

Noise

SpiDy.spectrum — Method

spectrum(n::ClassicalNoise)

Returns the classical noise at temperature n.T as a function of ω.

SpiDy.spectrum — Method

spectrum(n::QuantumNoise)

Returns the quantum noise at temperature n.T as a function of ω.

SpiDy.spectrum — Method

spectrum(n::NoZeroQuantumNoise)

Returns the quantum noise at temperature n.T as a function of ω. It differs from spectrum(n::QuantumNoise) in the fact that the zero-point fluctuations are removed.

Spectral density

SpiDy.sd — Method

sd(J::GenericSD)

Spectral density for generic shapes.

SpiDy.sdoverω — Method

sdoverω(J::GenericSD)

Spectral-density-divided-by-ω for generic shapes.

SpiDy.sdoverω — Method

sdoverω(J::LorentzianSD)

Spectral density divided by ω which naturally defines sd(J::LorentzianSD).

SpiDy.reorgenergy — Method

reorgenergy(J::GenericSD)

Reorganization energy numerically integrated as $\int_0^\infty \text{sdoverω}(\omega)d\omega$.

SpiDy.reorgenergy — Method

reorgenergy(J::LorentzianSD)

Analytical reorganization energy for Lorentzian spectral density.

SpiDy.kernel — Method

kernel(J::LorentzianSD)

Specific damping kernel for a Lorentzian spectral density defined by the parameters in J. It returns a function depending on ω.

SpiDy.psd — Method

psd(J::GenericSD, noise::Noise)

Power spectral density depending on parameters J and noise. It returns a function of ω.

SpiDy.psd — Method

psd(J::LorentzianSD, noise::ClassicalNoise)

Returns the analytical expression for power spectrum depending on Lorentzian spectral density and Classical noise.

Stochastic field

SpiDy.bfield — Method

bfield(N, Δt, J::GenericSD, noise::Noise, distro=nothing)

Returns the stochastic field b(t). It is evaluated using the Lorentzian spectral density defined by the parameters J, the classical/quantum/quantum-no-zero-energy noise. The sampling of the stochastic noise is done in frequency space. The default stochastic noise is white noise having Gaussian distribution but different distributions can be specified. N defines the number of steps and Δt defines the time step.

Coupling tensor

SpiDy.AnisoCoupling — Type

AnisoCoupling{TT<:AbstractMatrix{T} where {T<:Real}}

Returns an object of type Coupling built by passing it a 3x3 Matrix{Real} which defines the n-dimentional coupling between the spin and the stochastic fields. The matrix can therefore define a 1D, 2D as well as a 3D coupling.

SpiDy.IsoCoupling — Type

IsoCoupling{TT<:Real}

Returns an object of type Coupling built by passing it a single Real value which defines the n-dimentional isotropic coupling between the spin and the stochastic fields.

Dynamics

SpiDy.diffeqsolver — Method

diffeqsolver(s0, tspan, J::LorentzianSD, bfields, matrix::Coupling; S0=1/2, Bext=[0, 0, 1])

Returns [sol.t, s, sinterp], that is, the vector sol.t of time steps at which the solutions are evaluated, the 3-vector of the solutions s[1], s[2], s[3] evaluated at times sol.t, the 3 functions sinterp(t)[i] interpolations of the solutions s[i] found in the give time span.

The differential equation solver is built to account for Lorentzian spectral density. The two keyword arguments of the function are the spin length S0 set at default value 1/2 and the vector of the external magnetic field Bext set as unit-vector along the z-axis direction as default, Bext = [0, 0, 1].

Index

SpiDy.AnisoCoupling

SpiDy.IsoCoupling

SpiDy.bfield

SpiDy.diffeqsolver

SpiDy.kernel

SpiDy.psd

SpiDy.psd

SpiDy.reorgenergy

SpiDy.reorgenergy

SpiDy.sd

SpiDy.sdoverω

SpiDy.sdoverω

SpiDy.spectrum

SpiDy.spectrum

SpiDy.spectrum