SpiDy.jl documentation

SpiDy.jl documentation

Noise

Spectral density

Stochastic field

Coupling tensor

Dynamics

Index

Noise

SpiDy.Noise — Type

Noise

Definition of the abstract type Noise.

SpiDy.ClassicalNoise — Type

ClassicalNoise{TT<:Real}

Returns a ClassicalNoise structure of type Noise built by passing a Real value. This allows for future overloading of the spectrum method.

SpiDy.QuantumNoise — Type

ClassicalNoise{TT<:Real}

Returns a QuantumNoise structure of type Noise built by passing a Real value. This allows for future overloading of the spectrum method.

SpiDy.NoZeroQuantumNoise — Type

ClassicalNoise{TT<:Real}

Returns a NoZeroQuantumNoise structure of type Noise built by passing a Real value. This allows for future overloading of the spectrum method.

SpiDy.spectrum — Method

spectrum(n::ClassicalNoise)

Returns the classical noise at temperature n.T as a function of ω. The conditional statement makes sure the function does not divide by zero in case of ω==0.

SpiDy.spectrum — Method

spectrum(n::QuantumNoise)

Returns the quantum noise at temperature n.T as a function of ω. The conditional statement makes sure the function does not divide by zero in case of ω==0.

SpiDy.spectrum — Method

spectrum(n::NoZeroQuantumNoise)

Returns the quantum noise at temperature n.T as a function of ω. The conditional statement makes sure the function does not divide by zero in case of ω==0. It differs from spectrum(n::QuantumNoise) in the fact that the zero-point noise is removed.

Spectral density

SpiDy.GenericSD — Type

GenericSD

Definition of the abstract type GenericSD.

SpiDy.LorentzianSD — Type

LorentzianSD{T<:Real}

Returns a LorentzianSD structure of type GenericSD built by passing 3 Real values. The values are ordered as LorentzianSD(α, ω0, Γ).

Examples

julia> LorentzianSD(1., 3., 8.)

SpiDy.sd — Method

sd(J::GenericSD)

Defines the spectral density for generic shapes GenericSD.

SpiDy.sdoverω — Method

sdoverω(J::GenericSD)

Returns the spectral density divided by ω for generic shapes GenericSD.

SpiDy.sdoverω — Method

sdoverω(J::LorentzianSD)

Returns the spectral density divided by ω for LorentzianSD shapes which naturally defines sd(J::LorentzianSD).

SpiDy.reorgenergy — Method

reorgenergy(J::GenericSD)

Returns the reorganization energy numerically integrated as $\int_0^\infty \text{sdoverω}(\omega)d\omega$.

SpiDy.reorgenergy — Method

reorgenergy(J::LorentzianSD)

Returns the analytical reorganization energy for LorentzianSD shapes.

SpiDy.kernel — Method

kernel(J::LorentzianSD)

Returns the specific damping kernel for a Lorentzian spectral density defined by the parameters in J. It returns a function depending on ω.

SpiDy.psd — Method

psd(J::GenericSD, noise::Noise)

Returns the power spectral density depending on parameters J and noise. The returned function depends on ω.

SpiDy.psd — Method

psd(J::LorentzianSD, noise::ClassicalNoise)

Returns the analytical expression for power spectrum depending on Lorentzian spectral density and Classical noise.

Stochastic field

SpiDy.bfield — Method

bfield(N, Δt, J::GenericSD, noise::Noise, distro=nothing)

Returns the stochastic field b(t). It is evaluated using the Lorentzian spectral density defined by the parameters J, the classical/quantum/quantum-no-zero-energy noise. The sampling of the stochastic noise is done in frequency space. The default stochastic noise is white noise having Gaussian distribution but different distributions can be specified. N defines the number of steps and Δt defines the time step.

Coupling tensor

SpiDy.Coupling — Type

Coupling

Definition of the abstract type Coupling.

SpiDy.AnisoCoupling — Type

AnisoCoupling{TT<:AbstractMatrix{T} where {T<:Real}}

Returns a AnisoCoupling structure of type Coupling built by passing it a 3x3 Matrix{Real} which defines the n-dimentional coupling between the spin and the stochastic fields. The matrix can therefore define a 1D, 2D as well as a 3D coupling.

SpiDy.IsoCoupling — Type

IsoCoupling{TT<:Real}

Returns a IsoCoupling of type Coupling built by passing it a single Real value which defines the n-dimentional isotropic coupling between the spin and the stochastic fields.

Dynamics

SpiDy.diffeqsolver — Method

diffeqsolver(s0, tspan, J::LorentzianSD, bfields, matrix::Coupling; S0=1/2, Bext=[0, 0, 1], saveat=[])

Returns [sol.t, s, sinterp], that is, the vector sol.t of time steps at which the solutions are evaluated, the 3-vector of the solutions s[1], s[2], s[3] evaluated at times sol.t, the 3 functions sinterp(t)[i] interpolations of the solutions s[i] found in the give time span.

The differential equation solver is built to account for Lorentzian spectral density.

Keyword arguments:

S0 spin length set at default value 1/2, S0=1/2.

Bext external magnetic field set as unit-vector along the z-axis as default, Bext = [0, 0, 1]

saveat is an option of the function solve() which allows to only save the solution at the points needed to evaluate the steady-state,

i.e. at late times. Used to optimize memory management and speed of the solver. Default value is an empty list, saveat=[], resulting in the solution saved at optimal time steps withing the entire time span.

Examples

julia> diffeqsolver(s0, tspan, J, bfields, matrix; saveat=((N*4÷5):1:N)*Δt)

Index

SpiDy.AnisoCoupling

SpiDy.ClassicalNoise

SpiDy.Coupling

SpiDy.GenericSD

SpiDy.IsoCoupling

SpiDy.LorentzianSD

SpiDy.NoZeroQuantumNoise

SpiDy.Noise

SpiDy.QuantumNoise

SpiDy.bfield

SpiDy.diffeqsolver

SpiDy.kernel

SpiDy.psd

SpiDy.psd

SpiDy.reorgenergy

SpiDy.reorgenergy

SpiDy.sd

SpiDy.sdoverω

SpiDy.sdoverω

SpiDy.spectrum

SpiDy.spectrum

SpiDy.spectrum