Coupling tensor

SpiDy.Coupling — Type

Coupling

Definition of the abstract type Coupling.

SpiDy.AnisoCoupling — Type

AnisoCoupling{TT<:AbstractMatrix{T} where {T<:Real}}

Returns a AnisoCoupling structure of type Coupling built by passing it a 3x3 Matrix{Real} which defines the n-dimentional coupling between the spin and the stochastic fields. The matrix can therefore define a 1D, 2D as well as a 3D coupling.

SpiDy.IsoCoupling — Type

IsoCoupling{TT<:Real}

Returns a IsoCoupling of type Coupling built by passing it a single Real value which defines the n-dimentional isotropic coupling between the spin and the stochastic fields.