Noise

SpiDy.Noise — Type

abstract type Noise

An abstract type used to represent different kinds of environment noise.

Any user-defined subtype of Noise should implement a method noise(::NoiseSubType) which should return a function that represent the spectrum of the noise.

SpiDy.ClassicalNoise — Type

struct ClassicalNoise{TT<:Real} <: Noise

A subtype of Noise used to represent classical noise, with spectrum

\[\mathcal{S}(\omega) = \frac{k_\mathrm{B}T}{\hbar\omega}\]

Fields

T::TT: The temperature of the environment associated with the noise.

SpiDy.QuantumNoise — Type

struct QuantumNoise{TT<:Real} <: Noise

A subtype of Noise used to represent quantum noise, with spectrum

\[\mathcal{S}(\omega) = \coth\left(\frac{\hbar\omega}{2k_\mathrm{B}T}\right)\]

Fields

T::TT: The temperature of the environment associated with the noise.

SpiDy.NoZeroQuantumNoise — Type

struct NoZeroQuantumNoise{TT<:Real} <: Noise

A subtype of Noise used to represent quantum noise with zero-point-fluctuations removed, that is with spectrum

\[\mathcal{S}(\omega) = \coth\left(\frac{\hbar\omega}{2k_\mathrm{B}T}\right) - 1\]

Fields

T::TT: The temperature of the environment associated with the noise.

Missing docstring.

Missing docstring for spectrum(n::ClassicalNoise). Check Documenter's build log for details.

Missing docstring.

Missing docstring for spectrum(n::QuantumNoise). Check Documenter's build log for details.

Missing docstring.

Missing docstring for spectrum(n::NoZeroQuantumNoise). Check Documenter's build log for details.