Noise

SpiDy.Noise — Type

Noise

Definition of the abstract type Noise.

SpiDy.ClassicalNoise — Type

ClassicalNoise{TT<:Real}

Returns a ClassicalNoise structure of type Noise built by passing a Real value. This allows for future overloading of the spectrum method.

SpiDy.QuantumNoise — Type

ClassicalNoise{TT<:Real}

Returns a QuantumNoise structure of type Noise built by passing a Real value. This allows for future overloading of the spectrum method.

SpiDy.NoZeroQuantumNoise — Type

ClassicalNoise{TT<:Real}

Returns a NoZeroQuantumNoise structure of type Noise built by passing a Real value. This allows for future overloading of the spectrum method.

SpiDy.spectrum — Method

spectrum(n::ClassicalNoise)

Returns the classical noise at temperature n.T as a function of ω. The conditional statement makes sure the function does not divide by zero in case of ω==0.

SpiDy.spectrum — Method

spectrum(n::QuantumNoise)

Returns the quantum noise at temperature n.T as a function of ω. The conditional statement makes sure the function does not divide by zero in case of ω==0.

SpiDy.spectrum — Method

spectrum(n::NoZeroQuantumNoise)

Returns the quantum noise at temperature n.T as a function of ω. The conditional statement makes sure the function does not divide by zero in case of ω==0. It differs from spectrum(n::QuantumNoise) in the fact that the zero-point noise is removed.